Extract chart utility
